pyproject.toml
[project]
name = "proyecto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.6",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.0.6",
    "python-dotenv>=1.2.1",
]
